apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.6.5",
    "alembic>=1.18.4",
    "bcrypt>=4.2.0",
    "pyjwt>=2.10.0",
    "celery>=5.6.3",
    "docker>=7.0.0",
    "cryptography>=48.0.0",
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "langfuse>=4.5.1",
    "litellm>=1.83.0",
    "openinference-instrumentation-litellm>=0.1.30",
    "opentelemetry-api>=1.41.1",
    "opentelemetry-exporter-otlp>=1.41.1",
    "opentelemetry-exporter-otlp-proto-http>=1.41.1",
    "opentelemetry-sdk>=1.41.1",
    "psycopg2-binary>=2.9.12",
    "pydantic-settings>=2.14.0",
    "pygithub>=2.9.1",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "scalar-fastapi>=1.8.2",
    "sqlmodel>=0.0.38",
    "uvicorn[standard]>=0.46.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "ruff>=0.15.12",
    "ty>=0.0.34",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
